+++ Products/CMFPlone/browser/admin.py
@@ -6,6 +6,7 @@ from Products.CMFPlone.factory import _DEFAULT_PROFILE
 from Products.CMFPlone.factory import addPloneSite
 from Products.CMFPlone.interfaces import INonInstallable
 from Products.CMFPlone.interfaces import IPloneSiteRoot
+from Products.CMFPlone.utils import get_installer
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from Products.GenericSetup import BASE, EXTENSION
 from Products.GenericSetup import profile_registry
@@ -20,6 +21,7 @@ from urlparse import urljoin
 from urlparse import urlparse
 from zope.component import adapts
 from zope.component import getAllUtilitiesRegisteredFor
+from zope.component import getMultiAdapter
 from zope.component import getUtility
 from zope.component import queryMultiAdapter
 from zope.component import queryUtility
@@ -304,9 +306,10 @@ class Upgrade(BrowserView):
                 REQUEST=self.request,
                 dry_run=form.get('dry_run', False),
             )
-            qi = getattr(self.context, 'portal_quickinstaller')
-            pac_installed = qi.isProductInstalled('plone.app.contenttypes')
-            pac_installable = qi.isProductInstallable('plone.app.contenttypes')
+            qi = get_installer(self.context, self.request)
+            pac_installed = qi.is_product_installed('plone.app.contenttypes')
+            pac_installable = qi.is_product_installable(
+                'plone.app.contenttypes')
             advertise_dx_migration = pac_installable and not pac_installed
 
             return self.index(