+++ Products/CMFPlone/controlpanel/browser/quickinstaller.py
@@ -1,24 +1,511 @@
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone import PloneMessageFactory as _
 from Products.CMFPlone.interfaces import INonInstallable
+from Products.CMFQuickInstallerTool.interfaces import INonInstallable as \
+    QINonInstallable
 from Products.Five.browser import BrowserView
 from Products.GenericSetup import EXTENSION
+from Products.GenericSetup.tool import UNKNOWN
 from Products.statusmessages.interfaces import IStatusMessage
 from plone.memoize import view
 from zope.component import getAllUtilitiesRegisteredFor
 import logging
+import pkg_resources
+import transaction
+import warnings
+
+logger = logging.getLogger('Plone')
 
 
-class ManageProductsView(BrowserView):
+class InstallerView(BrowserView):
+    """View on all contexts for installing and uninstalling products.
+    """
+
+    def __init__(self, *args, **kwargs):
+        super(InstallerView, self).__init__(*args, **kwargs)
+        self.ps = getToolByName(self.context, 'portal_setup')
+        self.errors = {}
+
+    def listInstallableProducts(self, skipInstalled=True):
+        warnings.warn(
+            'listInstallableProducts is no longer supported since Plone 5.1. '
+            'It will be removed in Plone 6.0.',
+            DeprecationWarning)
+
+    def listInstalledProducts(self, skipInstalled=True):
+        warnings.warn(
+            'listInstalledProducts is no longer supported since Plone 5.1. '
+            'It will be removed in Plone 6.0.',
+            DeprecationWarning)
+
+    def is_profile_installed(self, profile_id):
+        return self.ps.getLastVersionForProfile(profile_id) != UNKNOWN
+
+    def is_product_installed(self, product_id):
+        profile = self.get_install_profile(product_id, allow_hidden=True)
+        if not profile:
+            return False
+        return self.is_profile_installed(profile['id'])
+
+    def isProductInstalled(self, product_name):
+        warnings.warn(
+            'isProductInstalled is deprecated since Plone 5.1. '
+            'It will be removed in Plone 6.0. '
+            'Use is_product_installed instead.',
+            DeprecationWarning)
+        return self.is_product_installed(product_name)
+
+    def notifyInstalled(self, productname, **kwargs):
+        warnings.warn(
+            'notifyInstalled is no longer supported since Plone 5.1. '
+            'It will be removed in Plone 6.0.',
+            DeprecationWarning)
+
+    def _install_profile_info(self, product_id):
+        """List extension profile infos of a given product.
+
+        From CMFQuickInstallerTool/QuickInstallerTool.py
+        _install_profile_info
+        """
+        profiles = self.ps.listProfileInfo()
+        # We are only interested in extension profiles for the product.
+        # TODO Remove the manual Products.* check here. It is still needed.
+        profiles = [
+            prof for prof in profiles
+            if prof['type'] == EXTENSION and (
+                prof['product'] == product_id or
+                prof['product'] == 'Products.%s' % product_id
+            )
+        ]
+        return profiles
+
+    def get_install_profiles(self, product_id):
+        """List all installer profile ids of the given name.
+
+        From CMFQuickInstallerTool/QuickInstallerTool.py
+        getInstallProfiles
+
+        TODO Might be superfluous.
+        """
+        return [prof['id'] for prof in self._install_profile_info(product_id)]
+
+    def _get_profile(self, product_id, name, strict=True, allow_hidden=False):
+        """Return profile with given name.
+
+        Also return None when no profiles are found at all.
+
+        :param product_id: id of product/package.
+            For example CMFPlone or plone.app.registry.
+        :type product_id: string
+        :param name: name of profile.
+            Usually 'default' or 'uninstall'.
+        :type name: string
+        :param strict: When True, return None when name is not found.
+            Otherwise fall back to the first profile.
+        :type strict: boolean
+        :param allow_hidden: Allow getting hidden profile.
+            A non hidden profile is always preferred.
+        :type allow_hidden: boolean
+        :returns: True on success, False otherwise.
+        :rtype: boolean
+        """
+        profiles = self._install_profile_info(product_id)
+        if not profiles:
+            return
+        utils = getAllUtilitiesRegisteredFor(INonInstallable)
+        hidden = []
+        for util in utils:
+            gnip = getattr(util, 'getNonInstallableProfiles', None)
+            if gnip is None:
+                continue
+            hidden.extend(gnip())
+
+        # We have prime candidates that we prefer, and have hidden candidates
+        # in case allow_hidden is True.
+        prime_candidates = []
+        hidden_candidates = []
+        for profile in profiles:
+            profile_id = profile['id']
+            profile_id_parts = profile_id.split(':')
+            if len(profile_id_parts) != 2:
+                logger.error("Profile with id '%s' is invalid." % profile_id)
+                continue
+            if allow_hidden and profile_id in hidden:
+                if profile_id_parts[1] == name:
+                    # This will especially be true for uninstall profiles,
+                    # which are usually hidden.
+                    return profile
+                hidden_candidates.append(profile)
+                continue
+            if profile_id_parts[1] == name:
+                return profile
+            prime_candidates.append(profile)
+        if strict:
+            return
+        if prime_candidates:
+            # QI used to pick the first profile.
+            # Return the first profile after all.
+            return prime_candidates[0]
+        if allow_hidden and hidden_candidates:
+            # Return the first hidden profile.
+            return hidden_candidates[0]
+
+    def get_install_profile(self, product_id, allow_hidden=False):
+        """Return the default install profile.
+
+        From CMFQuickInstallerTool/QuickInstallerTool.py
+        getInstallProfile
+
+        :param product_id: id of product/package
+        :type product_id: string
+        :param allow_hidden: Allow getting otherwise hidden profile.
+            In the UI this will be False, but you can set it to True in
+            for example a call from plone.app.upgrade where you want to
+            install a new core product, even though it is hidden for users.
+        :type allow_hidden: boolean
+        :returns: True on success, False otherwise.
+        :rtype: boolean
+        """
+        return self._get_profile(product_id, 'default', strict=False,
+                                 allow_hidden=allow_hidden)
+
+    def get_uninstall_profile(self, product_id):
+        """Return the uninstall profile.
+
+        Note: not used yet.
+        """
+        return self._get_profile(
+            product_id, 'uninstall', strict=True, allow_hidden=True)
+
+    def isProductInstallable(self, productname):
+        warnings.warn(
+            'isProductInstallable is deprecated since Plone 5.1. '
+            'It will be removed in Plone 6.0. '
+            'Use is_product_installable instead.',
+            DeprecationWarning)
+        return self.is_product_installable(productname)
+
+    def isProductAvailable(self, productname):
+        warnings.warn(
+            'isProductAvailable is deprecated since Plone 5.1. '
+            'It will be removed in Plone 6.0. '
+            'Use is_product_installable instead.',
+            DeprecationWarning)
+        return self.is_product_installable(productname)
+
+    def is_product_installable(self, product_id, allow_hidden=False):
+        """Does a product have an installation profile?
+
+        From CMFQuickInstallerTool/QuickInstallerTool.py
+        isProductInstallable (and the deprecated isProductAvailable)
+
+        :param allow_hidden: Allow installing otherwise hidden products.
+            In the UI this will be False, but you can set it to True in
+            for example a call from plone.app.upgrade where you want to
+            install a new core product, even though it is hidden for users.
+        :type allow_hidden: boolean
+        :returns: True when product is installable, False otherwise.
+        :rtype: boolean
+        """
+        if not allow_hidden:
+            not_installable = []
+            utils = getAllUtilitiesRegisteredFor(INonInstallable)
+            for util in utils:
+                gnip = getattr(util, 'getNonInstallableProducts', None)
+                if gnip is None:
+                    continue
+                not_installable.extend(gnip())
+            if product_id in not_installable:
+                return False
+            # BBB.  For backwards compatibility, we try the INonInstallable
+            # from the old QI as well.
+            not_installable = []
+            utils = getAllUtilitiesRegisteredFor(QINonInstallable)
+            for util in utils:
+                not_installable.extend(util.getNonInstallableProducts())
+            if product_id in not_installable:
+                return False
+
+        profile = self.get_install_profile(
+            product_id, allow_hidden=allow_hidden)
+        if profile is None:
+            return
+        try:
+            self.ps.getProfileDependencyChain(profile['id'])
+        except KeyError, e:
+            # Don't show twice the same error: old install and profile
+            # oldinstall is test in first in other methods we may have an extra
+            # 'Products.' in the namespace.
+            #
+            # TODO:
+            # 1. Make sense of the previous comment.
+            # 2. Possibly remove the special case for 'Products'.
+            # 3. Make sense of the next five lines: they remove 'Products.'
+            #    when it is there, and add it when it is not???
+            checkname = product_id
+            if checkname.startswith('Products.'):
+                checkname = checkname[9:]
+            else:
+                checkname = 'Products.' + checkname
+            if checkname in self.errors:
+                if self.errors[checkname]['value'] == e.args[0]:
+                    return False
+                # A new error is found, register it
+                self.errors[product_id] = dict(
+                    type=_(
+                        u"dependency_missing",
+                        default=u"Missing dependency"
+                    ),
+                    value=e.args[0],
+                    product_id=product_id
+                )
+            else:
+                self.errors[product_id] = dict(
+                    type=_(
+                        u"dependency_missing",
+                        default=u"Missing dependency"
+                    ),
+                    value=e.args[0],
+                    product_id=product_id
+                )
+            return False
+        return True
+
+    def getProductFile(self, productname, fname='readme.txt'):
+        warnings.warn(
+            'getProductFile is no longer supported since Plone 5.1. '
+            'It will be removed in Plone 6.0.',
+            DeprecationWarning)
+
+    def getProductReadme(self, productname):
+        warnings.warn(
+            'getProductReadme is no longer supported since Plone 5.1. '
+            'It will be removed in Plone 6.0.',
+            DeprecationWarning)
+
+    def getProductVersion(self, productname):
+        warnings.warn(
+            'getProductVersion is deprecated since Plone 5.1. '
+            'It will be removed in Plone 6.0. '
+            'Use get_product_version instead.',
+            DeprecationWarning)
+
+    def get_product_version(self, product_id):
+        """Return the version of the product (package).
+
+        From CMFQuickInstallerTool/QuickInstallerTool
+        getProductVersion
+        That implementation used to fall back to getting the version.txt.
+        """
+        try:
+            dist = pkg_resources.get_distribution(product_id)
+            return dist.version
+        except pkg_resources.DistributionNotFound:
+            return ''
+
+        # TODO: check if extra Products check is needed after all.
+        # if "." not in product_id:
+        #     try:
+        #         dist = pkg_resources.get_distribution(
+        #             "Products." + product_id)
+        #         return dist.version
+        #     except pkg_resources.DistributionNotFound:
+        #         pass
+
+    def get_latest_upgrade_step(self, profile_id):
+        """Get highest ordered upgrade step for profile.
+
+        If anything errors out then go back to "old way" by returning
+        'unknown'.
+
+        From CMFPlone/QuickInstallerTool.py getLatestUpgradeStep
+        """
+        profile_version = UNKNOWN
+        try:
+            available = self.ps.listUpgrades(profile_id, True)
+            if available:  # could return empty sequence
+                latest = available[-1]
+                profile_version = max(latest['dest'],
+                                      key=pkg_resources.parse_version)
+        except Exception:
+            pass
+        return profile_version
+
+    def upgrade_info(self, product_id):
+        """Returns upgrade info for a product.
+
+        This is a dict with among others two booleans values, stating if
+        an upgrade is required and available.
+
+        From CMFPlone/QuickInstaller.py upgradeInfo
+
+        :param product_id: id of product/package
+        :type product_id: string
+        :returns: dictionary with info about product
+        :rtype: dict
+        """
+        available = self.is_product_installable(product_id, allow_hidden=True)
+        if not available:
+            return {}
+        profile = self.get_install_profile(product_id, allow_hidden=True)
+        if profile is None:
+            # No GS profile, not supported.
+            return {}
+        profile_id = profile['id']
+        profile_version = str(self.ps.getVersionForProfile(profile_id))
+        if profile_version == 'latest':
+            profile_version = self.get_latest_upgrade_step(profile_id)
+        if profile_version == UNKNOWN:
+            # If a profile doesn't have a metadata.xml use the package version.
+            profile_version = self.get_product_version(product_id)
+        installed_profile_version = self.ps.getLastVersionForProfile(
+            profile_id)
+        # getLastVersionForProfile returns the version as a tuple or unknown.
+        if installed_profile_version != UNKNOWN:
+            installed_profile_version = str(
+                '.'.join(installed_profile_version))
+        return dict(
+            required=profile_version != installed_profile_version,
+            available=len(self.ps.listUpgrades(profile_id)) > 0,
+            hasProfile=True,  # TODO hasProfile is always True now.
+            installedVersion=installed_profile_version,
+            newVersion=profile_version,
+        )
+
+    def reinstallProducts(self, products, **kwargs):
+        """Reinstalls a list of products, the main difference to
+        uninstall/install is that it does not remove portal objects
+        created during install (e.g. tools, etc.)
+        """
+        warnings.warn(
+            'reinstallProducts is no longer supported since Plone 5.1. '
+            'It will be removed in Plone 6.0.',
+            DeprecationWarning)
+
+    def upgradeProduct(self, pid):
+        warnings.warn(
+            'upgradeProduct is deprecated since Plone 5.1. '
+            'It will be removed in Plone 6.0. '
+            'Use upgrade_product instead.',
+            DeprecationWarning)
+        return self.upgrade_product(pid)
+
+    def upgrade_product(self, product_id):
+        """Run the upgrade steps for a product.
+
+        Returns True on success, False otherwise.
+        """
+        profile = self.get_install_profile(product_id, allow_hidden=True)
+        if profile is None:
+            logger.error("Could not upgrade %s, no profile.", product_id)
+            return False
+        self.ps.upgradeProfile(profile['id'])
+        return True
+
+    def installProducts(self, products=None, **kwargs):
+        warnings.warn(
+            'installProducts is deprecated since Plone 5.1. '
+            'It will be removed in Plone 6.0. '
+            'Use install_product with a single product instead. '
+            'All keyword arguments are ignored.',
+            DeprecationWarning)
+        if products is None:
+            products = []
+        for product_id in products:
+            self.install_product(product_id)
+
+    def installProduct(self, product_name, **kwargs):
+        """Deprecated install product.
+        """
+        warnings.warn(
+            'installProduct is deprecated since Plone 5.1. '
+            'It will be removed in Plone 6.0. '
+            'Use install_product instead. '
+            'All keyword arguments are ignored.',
+            DeprecationWarning)
+        return self.install_product(product_name)
+
+    def install_product(self, product_id, allow_hidden=False):
+        """Install a product by name.
+
+        From CMFQuickInstallerTool/QuickInstallerTool.py installProduct
+
+        :param product_id: id of product/package
+        :type product_id: string
+        :param allow_hidden: Allow installing otherwise hidden products.
+            In the UI this will be False, but you can set it to True in
+            for example a call from plone.app.upgrade where you want to
+            install a new core product, even though it is hidden for users.
+        :type allow_hidden: boolean
+        :returns: True on success, False otherwise.
+        :rtype: boolean
+        """
+        profile = self.get_install_profile(
+            product_id, allow_hidden=allow_hidden)
+        if not profile:
+            logger.error("Could not install %s: no profile found.", product_id)
+            # TODO Possibly raise an error.
+            return False
+
+        if self.is_product_installed(product_id):
+            logger.error("Could not install %s: profile already installed.",
+                         product_id)
+            return False
+
+        # Okay, actually install the profile.
+        profile_id = profile['id']
+        self.ps.runAllImportStepsFromProfile('profile-%s' % profile_id)
+
+        if not self.is_profile_installed(profile_id):
+            version = self.get_product_version(product_id)
+            logger.warn('Profile %s has no metadata.xml version. Falling back '
+                        'to package version %s', profile_id, version)
+            self.ps.setLastVersionForProfile(profile_id, version)
+
+        # No problems encountered.
+        return True
+
+    def uninstallProducts(self, products=None, **kwargs):
+        """Uninstall a list of products.
+        """
+        warnings.warn(
+            'uninstallProducts is deprecated since Plone 5.1. '
+            'It will be removed in Plone 6.0. '
+            'Use uninstall_product with a single product instead. '
+            'All keyword arguments are ignored.',
+            DeprecationWarning)
+        if products is None:
+            products = []
+        for product_id in products:
+            self.uninstall_product(product_id)
+
+    def uninstall_product(self, product_id):
+        """Uninstall a product by name.
+
+        Returns True on success, False otherwise.
+        """
+        profile = self.get_uninstall_profile(product_id)
+        if not profile:
+            logger.error("Could not uninstall %s: no uninstall profile "
+                         "found.", product_id)
+            return False
+
+        self.ps.runAllImportStepsFromProfile(
+            'profile-%s' % profile['id'])
+
+        # Unmark the install profile.
+        install_profile = self.get_install_profile(
+            product_id, allow_hidden=True)
+        if install_profile:
+            self.ps.unsetLastVersionForProfile(install_profile['id'])
+        return True
+
+
+class ManageProductsView(InstallerView):
     """
     Activate and deactivate products in mass, and without weird
     permissions issues
     """
-
-    def __init__(self, *args, **kwargs):
-        super(ManageProductsView, self).__init__(*args, **kwargs)
-        self.qi = getToolByName(self.context, 'portal_quickinstaller')
-        self.ps = getToolByName(self.context, 'portal_setup')
 
     def __call__(self):
         return self.index()
@@ -28,11 +515,21 @@ class ManageProductsView(BrowserView):
         addons = {}
 
         ignore_profiles = []
+        ignore_products = []
         utils = getAllUtilitiesRegisteredFor(INonInstallable)
         for util in utils:
-            ignore_profiles.extend(util.getNonInstallableProfiles())
+            ni_profiles = getattr(util, 'getNonInstallableProfiles', None)
+            if ni_profiles is not None:
+                ignore_profiles.extend(ni_profiles())
+            ni_products = getattr(util, 'getNonInstallableProducts', None)
+            if ni_products is not None:
+                ignore_products.extend(ni_products())
 
+        # Known profiles:
         profiles = self.ps.listProfileInfo()
+        # Profiles that have upgrade steps (which may or may not have been
+        # applied already).
+        # profiles_with_upgrades = self.ps.listProfilesWithUpgrades()
         for profile in profiles:
             if profile['type'] != EXTENSION:
                 continue
@@ -42,66 +539,58 @@ class ManageProductsView(BrowserView):
                 continue
             pid_parts = pid.split(':')
             if len(pid_parts) != 2:
-                logging.error("Profile with id '%s' is invalid." % pid)
+                logger.error("Profile with id '%s' is invalid." % pid)
+            # Which package (product) is this from?
             product_id = profile['product']
+            if product_id in ignore_products:
+                continue
             profile_type = pid_parts[-1]
             if product_id not in addons:
                 # get some basic information on the product
-                product_file = self.qi.getProductFile(product_id)
-                installed = False
-                upgrade_info = None
-                p_obj = self.qi._getOb(product_id, None)
-                if p_obj:
-                    # TODO; if you install then uninstall, the
-                    # presence lingers in the qi. Before it is
-                    # run the very first time, it doesn't exist
-                    # at all in the qi. How remove the qi from this?
-                    installed = p_obj.isInstalled()
-                    upgrade_info = self.qi.upgradeInfo(product_id)
-                else:
-                    # XXX: holy rabbit hole batman!
-                    if not self.qi.isProductInstallable(product_id):
-                        continue
-
-                if profile_type in product_id:
-                    profile_type = 'default'
-                    # XXX override here so some products that do not
-                    # explicitly say "default" for their install
-                    # profile still work
-                    # I'm not sure this is right but this is a way
-                    # to get CMFPlacefulWorkflow to show up in addons
-                    # If it's safe to rename profiles, we can do that too
-
+                installed = self.is_profile_installed(pid)
+                upgrade_info = {}
+                if installed:
+                    upgrade_info = self.upgrade_info(product_id)
+                elif not self.is_product_installable(product_id):
+                    continue
                 addons[product_id] = {
                     'id': product_id,
                     'title': product_id,
                     'description': '',
-                    'product_file': product_file,
                     'upgrade_profiles': {},
                     'other_profiles': [],
                     'install_profile': None,
+                    'install_profile_id': '',
                     'uninstall_profile': None,
+                    'uninstall_profile_id': '',
                     'is_installed': installed,
                     'upgrade_info': upgrade_info,
                     'profile_type': profile_type,
                 }
-            product = addons[product_id]
-            if profile_type == 'default':
-                product['title'] = profile['title']
-                product['description'] = profile['description']
-                product['install_profile'] = profile
-                product['profile_type'] = profile_type
-            elif profile_type == 'uninstall':
-                product['uninstall_profile'] = profile
-                if 'profile_type' not in product:
-                    # if this is the only profile installed, it could just be an uninstall
-                    # profile
-                    product['profile_type'] = profile_type
+                # Add info on install and uninstall profile.
+                product = addons[product_id]
+                install_profile = self.get_install_profile(product_id)
+                if install_profile is not None:
+                    product['title'] = install_profile['title']
+                    product['description'] = install_profile['description']
+                    product['install_profile'] = install_profile
+                    product['install_profile_id'] = install_profile['id']
+                    product['profile_type'] = 'default'
+                uninstall_profile = self.get_uninstall_profile(product_id)
+                if uninstall_profile is not None:
+                    product['uninstall_profile'] = uninstall_profile
+                    product['uninstall_profile_id'] = uninstall_profile['id']
+                    # Do not override profile_type.
+                    if not product['profile_type']:
+                        product['profile_type'] = 'uninstall'
+            if profile['id'] in (product['install_profile_id'],
+                                 product['uninstall_profile_id']):
+                # Everything has been done.
+                continue
+            elif 'version' in profile:
+                product['upgrade_profiles'][profile['version']] = profile
             else:
-                if 'version' in profile:
-                    product['upgrade_profiles'][profile['version']] = profile
-                else:
-                    product['other_profiles'].append(profile)
+                product['other_profiles'].append(profile)
         return addons
 
     def get_addons(self, apply_filter=None, product_name=None):
@@ -109,7 +598,7 @@ class ManageProductsView(BrowserView):
         100% based on generic setup profiles now. Kinda.
         For products magic, use the zope quickinstaller I guess.
 
-        @filter:= 'installed': only products that are installed
+        @filter:= 'installed': only products that are installed and not hidden
                   'upgrades': only products with upgrades
                   'available': products that are not installed bit
                                could be
@@ -124,7 +613,7 @@ class ManageProductsView(BrowserView):
         addons = self.marshall_addons()
         filtered = {}
         if apply_filter == 'broken':
-            all_broken = self.qi.getBrokenInstalls()
+            all_broken = self.errors.values()
             for broken in all_broken:
                 filtered[broken['productname']] = broken
         else:
@@ -142,12 +631,8 @@ class ManageProductsView(BrowserView):
                     if addon['profile_type'] != 'default':
                         continue
                 elif apply_filter == 'upgrades':
-                    # weird p.a.discussion integration behavior
                     upgrade_info = addon['upgrade_info']
-                    if type(upgrade_info) == bool:
-                        continue
-
-                    if not upgrade_info['available']:
+                    if not upgrade_info.get('available'):
                         continue
 
                 filtered[product_id] = addon
@@ -169,75 +654,80 @@ class ManageProductsView(BrowserView):
     def get_broken(self):
         return self.get_addons(apply_filter='broken').values()
 
-    def upgrade_product(self, product):
-        qi = getToolByName(self.context, 'portal_quickinstaller')
-        messages = IStatusMessage(self.request)
-        try:
-            qi.upgradeProduct(product)
-            messages.addStatusMessage(
-                _(u'Upgraded ${product}!', mapping={'product': product}), type="info")
-            return True
-        except Exception, e:
-            logging.error("Could not upgrade %s: %s" % (product, e))
-            messages.addStatusMessage(
-                _(u'Error upgrading ${product}.', mapping={'product': product}), type="error")
 
-        return False
-
-
-class UpgradeProductsView(BrowserView):
+class UpgradeProductsView(InstallerView):
     """
     Upgrade a product... or twenty
     """
 
     def __call__(self):
-        qi = ManageProductsView(self.context, self.request)
         products = self.request.get('prefs_reinstallProducts', None)
         if products:
-            for product in products:
-                qi.upgrade_product(product)
+            messages = IStatusMessage(self.request)
+            for product_id in products:
+                result = self.upgrade_product(product_id)
+                if not result:
+                    messages.addStatusMessage(
+                        _(u'Error upgrading ${product}.',
+                          mapping={'product': product_id}), type="error")
+                    # Abort changes for all upgrades.
+                    transaction.abort()
+                    break
+            else:
+                messages.addStatusMessage(
+                    _(u'Upgraded products.'), type="info")
 
         purl = getToolByName(self.context, 'portal_url')()
         self.request.response.redirect(purl + '/prefs_install_products_form')
 
 
-class InstallProductsView(BrowserView):
+class InstallProductsView(InstallerView):
 
     def __call__(self):
-        qi = getToolByName(self.context, 'portal_quickinstaller')
-        products = self.request.get('install_products')
-        msg_type = 'info'
-        if products:
+        product_id = self.request.get('install_product')
+        if product_id:
             messages = IStatusMessage(self.request)
-            for product in products:
-                qi.installProducts(products=[product, ])
-                msg = _(u'Installed ${product}!', mapping={'product': product})
-                messages.addStatusMessage(msg, type=msg_type)
+            msg_type = 'info'
+            result = self.install_product(product_id)
+            if result:
+                msg = _(u'Installed ${product}!',
+                        mapping={'product': product_id})
+            else:
+                # Only reason should be that between loading the page and
+                # clicking to install a product, another user has already
+                # installed this product.
+                msg_type = 'error'
+                msg = _(u'Failed to install ${product}.',
+                        mapping={'product': product_id})
+            messages.addStatusMessage(msg, type=msg_type)
 
         purl = getToolByName(self.context, 'portal_url')()
         self.request.response.redirect(purl + '/prefs_install_products_form')
 
 
-class UninstallProductsView(BrowserView):
+class UninstallProductsView(InstallerView):
 
     def __call__(self):
-        qi = getToolByName(self.context, 'portal_quickinstaller')
-        products = self.request.get('uninstall_products')
-        msg_type = 'info'
-        if products:
+        product_id = self.request.get('uninstall_product')
+        if product_id:
             messages = IStatusMessage(self.request)
-            # 1 at a time for better error messages
-            for product in products:
-                try:
-                    qi.uninstallProducts(products=[product, ])
+            try:
+                result = self.uninstall_product(product_id)
+            except Exception, e:
+                logger.error("Could not uninstall %s: %s", product_id, e)
+                msg_type = 'error'
+                msg = _(u'Error uninstalling ${product}.', mapping={
+                        'product': product_id})
+            else:
+                if result:
+                    msg_type = 'info'
                     msg = _(u'Uninstalled ${product}.',
-                            mapping={'product': product})
-                except Exception, e:
-                    logging.error("Could not uninstall %s: %s" % (product, e))
+                            mapping={'product': product_id})
+                else:
                     msg_type = 'error'
-                    msg = _(u'Error uninstalling ${product}.', mapping={
-                            'product': product})
-                messages.addStatusMessage(msg, type=msg_type)
+                    msg = _(u'Could not uninstall ${product}.',
+                            mapping={'product': product_id})
+            messages.addStatusMessage(msg, type=msg_type)
 
         purl = getToolByName(self.context, 'portal_url')()
         self.request.response.redirect(purl + '/prefs_install_products_form')